Refactor/codebase

+++ src/api/usecases/task/delete.usecase.ts
@@ -1,15 +1,20 @@
 import { Task } from "@prisma/client";
 import { TaskRepository } from "../../repositories/task.repository";
+import { NotFoundError } from "../../errors/AppError";
 
 type UseCaseParam = {
   id: number;
-  userId: number
+  userId: number;
 };
 
 export class DeleteTaskUseCase {
   constructor(private taskRepository: TaskRepository) {}
 
   async execute({ id, userId }: UseCaseParam): Promise<Task> {
-    return this.taskRepository.delete(id, userId);
+    if (!id) {
+      throw new NotFoundError("Task not found");
+    } else {
+      return this.taskRepository.delete(id, userId);
+    }
   }
 }
